Create Alphabet_randomizer
A little Python code to randomize the letters of the Alphabet. This script is written in Python 3.4 by Artengar, date May 2015.
It was made with influences out of presentations performed by Olivier Ibsen. This is a group of two graphic designers who base their presentations on a letter they get from a little script. According to the letter delivered by the script, they showed/talked about different subjects.